+++ src/shared/records.rs
@@ -1,2 +1,2 @@
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Affordance(pub(crate) u32);
+pub struct Affordance(pub(crate) u64);
